skills/packs/video-production/requirements.txt
librosa>=0.10.0
Pillow>=10.0.0
python-dotenv>=1.0.0
openai-whisper>=20231117
requests>=2.28.0
yt-dlp>=2024.0.0
